+++ CMakeLists.txt
@@ -16,4 +16,4 @@ fprime_setup_included_code()
 
 
 # This includes project-wide objects
-include("${CMAKE_CURRENT_LIST_DIR}/project.cmake")
+add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/LedBlinker")
