+++ LedBlinker/CMakeLists.txt
@@ -1,28 +1,5 @@
-#####
-# 'LedBlinker' Deployment:
-#
-# This registers the 'LedBlinker' deployment to the build system. 
-# Custom components that have not been added at the project-level should be added to 
-# the list below.
-#
-#####
+# This CMake file is intended to register project-wide objects.
+# This allows for reuse between deployments, or other projects.
 
-# Deployment is only compatible with the Arduino platform
-restrict_platforms(ArduinoFw)
-
-###
-# Topology and Components
-###
-add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/config/")
-add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Top/")
-
-# Add custom components to this specific deployment here
-# add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/MyComponent/")
-
-
-set(SOURCE_FILES "${CMAKE_CURRENT_LIST_DIR}/Main.cpp")
-set(MOD_DEPS ${FPRIME_CURRENT_MODULE}/Top)
-
-register_fprime_deployment()
-
-finalize_arduino_executable("${FPRIME_CURRENT_MODULE}")
+add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Components")
+add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/LedBlinkerDeployment/")
