+++ LedBlinker/Components/Led/Led.fpp
@@ -1,4 +1,4 @@
-module Components {
+module LedBlinker {
     @ Component to blink an LED driven by a rate group
     active component Led {
