+++ LedBlinker/LedBlinkerDeployment/Top/LedBlinkerDeploymentTopology.cpp
@@ -1,19 +1,20 @@
 // ======================================================================
-// \title  LedBlinkerTopology.cpp
+// \title  LedBlinkerDeploymentTopology.cpp
 // \brief cpp file containing the topology instantiation code
 //
 // ======================================================================
 // Provides access to autocoded functions
-#include <LedBlinker/Top/LedBlinkerTopologyAc.hpp>
+#include <LedBlinker/LedBlinkerDeployment/Top/LedBlinkerDeploymentTopologyAc.hpp>
 // Note: Uncomment when using Svc:TlmPacketizer
-// #include <LedBlinker/Top/LedBlinkerPacketsAc.hpp>
-#include <config/FppConstantsAc.hpp>
+// #include <LedBlinkerDeployment/Top/LedBlinkerDeploymentPacketsAc.hpp>
+#include <Fw/Logger/Logger.hpp>
 
 // Necessary project-specified types
 #include <Arduino/config/FprimeArduino.hpp>
 
-// Allows easy reference to objects in FPP/autocoder required namespaces
-using namespace LedBlinker;
+// Public functions for use in main program are namespaced with deployment module LedBlinker
+// This is also the namespace where the topology components are instantiated by FPP.
+namespace LedBlinker {
 
 // The reference topology divides the incoming clock signal (1Hz) into sub-signals: 1/100Hz, 1/200Hz, and 1/1000Hz
 Svc::RateGroupDriver::DividerSet rateGroupDivisors{{{100, 0}, {200, 0}, {1000, 0}}};
@@ -21,6 +22,8 @@ Svc::RateGroupDriver::DividerSet rateGroupDivisors{{{100, 0}, {200, 0}, {1000, 0
 // Rate groups may supply a context token to each of the attached children whose purpose is set by the project. The
 // reference topology sets each token to zero as these contexts are unused in this project.
 U32 rateGroup1Context[FppConstant_PassiveRateGroupOutputPorts::PassiveRateGroupOutputPorts] = {};
+
+
 
 /**
  * \brief configure/setup components in project-specific way
@@ -39,8 +42,6 @@ void configureTopology() {
     gpioDriver.open(Arduino::DEF_LED_BUILTIN, Arduino::GpioDriver::GpioDirection::OUT);
 }
 
-// Public functions for use in main program are namespaced with deployment name LedBlinker
-namespace LedBlinker {
 void setupTopology(const TopologyState& state) {
     // Autocoded initialization. Function provided by autocoder.
     initComponents(state);
@@ -59,10 +60,17 @@ void setupTopology(const TopologyState& state) {
     // loadParameters();
     // Autocoded task kick-off (active components). Function provided by autocoder.
     startTasks(state);
-    
-    rateDriver.configure(1);
+
     comDriver.configure(&Serial);
+}
+
+void startRateGroups(const Fw::TimeInterval& interval) {
+    rateDriver.configure(1); // TODO: change HardwareRateDriver to take Fw::TimeInterval
     rateDriver.start();
+}
+
+void stopRateGroups() {
+    rateDriver.stop();
 }
 
 void teardownTopology(const TopologyState& state) {
@@ -70,4 +78,4 @@ void teardownTopology(const TopologyState& state) {
     stopTasks(state);
     freeThreads(state);
 }
-};  // namespace LedBlinker
+};  // namespace LedBlinkerDeployment
