+++ LedBlinker/LedBlinkerDeployment/config/TlmChanImplCfg.hpp
@@ -46,7 +46,6 @@ namespace {
                                         // Works best when set to about twice the number of components producing telemetry
         TLMCHAN_HASH_MOD_VALUE = 99,    // !< The modulo value of the hashing function.
                                         // Should be set to a little below the ID gaps to spread the entries around
-
         TLMCHAN_HASH_BUCKETS = 18       // !< Buckets assignable to a hash slot.
                                         // Buckets must be >= number of telemetry channels in system
     };
