--- LedBlinker/README.md
@@ -1,28 +1,0 @@
-# LedBlinker Application
-
-This deployment was auto-generated by the F' utility tool.
-
-## Building and Running the LedBlinker Application
-
-In order to build the LedBlinker application, or any other F´ application, we first need to generate a build directory. This can be done with the following commands:
-
-```
-fprime-util generate
-```
-
-The next step is to build the LedBlinker application's code.
-```
-fprime-util build
-```
-
-## Running the application and F' GDS
-
-The following command will spin up the F' GDS as well as run the application binary and the components necessary for the GDS and application to communicate.
-
-```
-fprime-gds -n --dictionary ./build-artifacts/<build name>/LedBlinker/dict/LedBlinkerTopologyAppDictionary.xml --communication-selection uart --uart-device /dev/ttyACM0 --uart-baud 115200
-```
-> Change `<build name>` to the build of your deployment (i.e. `teensy41`, `featherM0`, etc.).
-
-> `/dev/ttyACM0` may vary for your system/device. It may also be `/dev/ttyUSB0`. For MacOS, it will be along the lines of `/dev/tty.usbmodem12345`. Change accordingly.
-> To view the list of your connected devices, run: `ls /dev/tty*`.
