--- LedBlinker/Top/CMakeLists.txt
@@ -1,21 +1,0 @@
-####
-# F prime CMakeLists.txt:
-#
-# SOURCE_FILES: combined list of source and autocoding files
-# MOD_DEPS: (optional) module dependencies
-####
-
-set(SOURCE_FILES
-  "${CMAKE_CURRENT_LIST_DIR}/instances.fpp"
-  # Note: Uncomment when using Svc:TlmPacketizer
-  # "${CMAKE_CURRENT_LIST_DIR}/LedBlinkerPackets.xml"
-  "${CMAKE_CURRENT_LIST_DIR}/topology.fpp"
-  "${CMAKE_CURRENT_LIST_DIR}/LedBlinkerTopology.cpp"
-)
-set(MOD_DEPS
-  Fw/Logger
-  Arduino/Drv/StreamDriver
-  fprime-baremetal/Os/TaskRunner
-)
-
-register_fprime_module()
