+++ settings.ini
@@ -5,7 +5,7 @@ library_locations: ./lib/fprime-arduino:./lib/fprime-baremetal
 
 default_toolchain: teensy41
 
-deployment_cookiecutter: https://github.com/fprime-community/fprime-arduino-deployment-cookiecutter.git 
+deployment_cookiecutter: https://github.com/ethancheez/fprime-arduino-deployment-cookiecutter.git 
 
 default_cmake_options:  FPRIME_ENABLE_FRAMEWORK_UTS=OFF
                         FPRIME_ENABLE_AUTOCODER_UTS=OFF
